+++ frontend/src/components/files/ListingItem.vue
@@ -210,8 +210,13 @@ const drop = async (event: Event) => {
   action(overwrite, rename);
 };
 
-const itemClick = (event: Event) => {
-  if (singleClick.value && !fileStore.multiple) open();
+const itemClick = (event: Event | KeyboardEvent) => {
+  if (
+    !((event as KeyboardEvent).ctrlKey || (event as KeyboardEvent).metaKey) &&
+    singleClick.value &&
+    !fileStore.multiple
+  )
+    open();
   else click(event);
 };