+++ frontend/src/components/prompts/Prompts.vue
@@ -26,6 +26,7 @@ import ReplaceRename from "./ReplaceRename.vue";
 import Share from "./Share.vue";
 import ShareDelete from "./ShareDelete.vue";
 import Upload from "./Upload.vue";
+import DiscardEditorChanges from "./DiscardEditorChanges.vue";
 
 const layoutStore = useLayoutStore();
 
@@ -49,6 +50,7 @@ const components = new Map<string, any>([
   ["upload", Upload],
   ["share-delete", ShareDelete],
   ["deleteUser", DeleteUser],
+  ["discardEditorChanges", DiscardEditorChanges],
 ]);
 
 watch(currentPromptName, (newValue) => {