+++ frontend/src/views/Share.vue
@@ -29,7 +29,7 @@
     <breadcrumbs :base="'/share/' + hash" />
 
     <div v-if="layoutStore.loading">
-      <h2 class="message delayed">
+      <h2 class="message delayed" style="padding-top: 3em !important">
         <div class="spinner">
           <div class="bounce1"></div>
           <div class="bounce2"></div>
@@ -75,28 +75,44 @@
     </div>
     <div v-else-if="req !== null">
       <div class="share">
-        <div class="share__box share__box__info">
-          <div class="share__box__header">
+        <div
+          class="share__box share__box__info"
+          style="
+            position: -webkit-sticky;
+            position: sticky;
+            top: -20.6em;
+            z-index: 999;
+          "
+        >
+          <div class="share__box__header" style="height: 3em">
             {{
               req.isDir
                 ? t("download.downloadFolder")
                 : t("download.downloadFile")
             }}
           </div>
-          <div class="share__box__element share__box__center share__box__icon">
+          <div
+            v-if="!req.isDir"
+            class="share__box__element share__box__center share__box__icon"
+          >
             <i class="material-icons">{{ icon }}</i>
           </div>
-          <div class="share__box__element">
-            <strong>{{ t("prompts.displayName") }}</strong> {{ req.name }}
+          <div class="share__box__element" style="height: 3em">
+            <strong>{{ $t("prompts.displayName") }}</strong> {{ req.name }}
           </div>
-          <div class="share__box__element" :data-title="modTime">
-            <strong>{{ t("prompts.lastModified") }}:</strong> {{ humanTime }}
+          <div v-if="!req.isDir" class="share__box__element" :title="modTime">
+            <strong>{{ $t("prompts.lastModified") }}:</strong> {{ humanTime }}
           </div>
-          <div class="share__box__element">
-            <strong>{{ t("prompts.size") }}:</strong> {{ humanSize }}
+          <div class="share__box__element" style="height: 3em">
+            <strong>{{ $t("prompts.size") }}:</strong> {{ humanSize }}
           </div>
           <div class="share__box__element share__box__center">
-            <a target="_blank" :href="link" class="button button--flat">
+            <a
+              target="_blank"
+              :href="link"
+              class="button button--flat"
+              style="height: 4em"
+            >
               <div>
                 <i class="material-icons">file_download</i
                 >{{ t("buttons.download") }}
@@ -113,12 +129,111 @@
                 >{{ t("buttons.openFile") }}
               </div>
             </a>
+            <qrcode-vue
+              v-if="req.isDir"
+              :value="link"
+              :size="100"
+              level="M"
+            ></qrcode-vue>
           </div>
-          <div class="share__box__element share__box__center">
+          <div v-if="!req.isDir" class="share__box__element share__box__center">
             <qrcode-vue :value="link" :size="200" level="M"></qrcode-vue>
+          </div>
+          <div
+            v-if="req.isDir"
+            class="share__box__element share__box__header"
+            style="height: 3em"
+          >
+            {{ $t("sidebar.preview") }}
+          </div>
+          <div
+            v-if="req.isDir"
+            class="share__box__element share__box__center share__box__icon"
+            style="padding: 0em !important; height: 12em !important"
+          >
+            <a
+              target="_blank"
+              :href="raw"
+              class="button button--flat"
+              v-if="
+                !fileStore.multiple &&
+                fileStore.selectedCount === 1 &&
+                req.items[fileStore.selected[0]].type === 'image'
+              "
+              style="height: 12em; padding: 0; margin: 0"
+            >
+              <img style="height: 12em" :src="raw" />
+            </a>
+            <div
+              v-else-if="
+                fileStore.multiple &&
+                fileStore.selectedCount === 1 &&
+                req.items[fileStore.selected[0]].type === 'audio'
+              "
+              style="height: 12em; padding-top: 1em; margin: 0"
+            >
+              <button
+                @click="play"
+                v-if="!tag"
+                style="
+                  font-size: 6em !important;
+                  border: 0px;
+                  outline: none;
+                  background: white;
+                "
+                class="material-icons"
+              >
+                play_circle_filled
+              </button>
+              <button
+                @click="play"
+                v-if="tag"
+                style="
+                  font-size: 6em !important;
+                  border: 0px;
+                  outline: none;
+                  background: white;
+                "
+                class="material-icons"
+              >
+                pause_circle_filled
+              </button>
+              <audio
+                id="myaudio"
+                ref="audio"
+                :src="raw"
+                controls
+                :autoplay="tag"
+              ></audio>
+            </div>
+            <video
+              v-else-if="
+                !fileStore.multiple &&
+                fileStore.selectedCount === 1 &&
+                req.items[fileStore.selected[0]].type === 'video'
+              "
+              style="height: 12em; padding: 0; margin: 0"
+              :src="raw"
+              controls
+            >
+              Sorry, your browser doesn't support embedded videos, but don't
+              worry, you can <a :href="raw">download it</a>
+              and watch it with your favorite video player!
+            </video>
+            <i
+              v-else-if="
+                !fileStore.multiple &&
+                fileStore.selectedCount === 1 &&
+                req.items[fileStore.selected[0]].isDir
+              "
+              class="material-icons"
+              >folder
+            </i>
+            <i v-else class="material-icons">call_to_action</i>
           </div>
         </div>
         <div
+          id="shareList"
           v-if="req.isDir && req.items.length > 0"
           class="share__box share__box__items"
         >
@@ -195,7 +310,7 @@ import QrcodeVue from "qrcode.vue";
 import Item from "@/components/files/ListingItem.vue";
 import { useFileStore } from "@/stores/file";
 import { useLayoutStore } from "@/stores/layout";
-import { computed, inject, onMounted, ref, watch } from "vue";
+import { computed, inject, onMounted, onBeforeUnmount, ref, watch } from "vue";
 import { useRoute } from "vue-router";
 import { useI18n } from "vue-i18n";
 import { StatusError } from "@/api/utils";
@@ -207,6 +322,8 @@ const password = ref<string>("");
 const attemptedPasswordLogin = ref<boolean>(false);
 const hash = ref<string>("");
 const token = ref<string>("");
+const audio = ref<HTMLAudioElement>();
+const tag = ref<boolean>(false);
 
 const $showSuccess = inject<IToastSuccess>("$showSuccess")!;
 
@@ -235,6 +352,16 @@ const icon = computed(() => {
 });
 
 const link = computed(() => (req.value ? api.getDownloadURL(req.value) : ""));
+const raw = computed(() => {
+  return req.value
+    ? req.value.items[fileStore.selected[0]].url.replace(
+        /share/,
+        "api/public/dl"
+      ) +
+        "?token=" +
+        token.value
+    : "";
+});
 const inlineLink = computed(() =>
   req.value ? api.getDownloadURL(req.value, true) : ""
 );
@@ -251,11 +378,20 @@ const humanTime = computed(() => dayjs(req.value?.modified).fromNow());
 const modTime = computed(() =>
   req.value
     ? new Date(Date.parse(req.value.modified)).toLocaleString()
-    : new Date()
+    : new Date().toLocaleString()
 );
 
 // Functions
 const base64 = (name: any) => Base64.encodeURI(name);
+const play = () => {
+  if (tag.value) {
+    audio.value?.pause();
+    tag.value = false;
+  } else {
+    audio.value?.play();
+    tag.value = true;
+  }
+};
 const fetchData = async () => {
   fileStore.reload = false;
   fileStore.selected = [];
@@ -366,6 +502,29 @@ const copyToClipboard = (text: string) => {
 onMounted(async () => {
   // Created
   hash.value = route.params.path[0];
+  window.addEventListener("keydown", keyEvent);
   await fetchData();
 });
+
+onBeforeUnmount(() => {
+  // Destroyed
+  window.removeEventListener("keydown", keyEvent);
+});
 </script>
+
+<style scoped>
+#listing.list {
+  height: auto;
+}
+
+#shareList {
+  overflow-y: scroll;
+}
+
+@media (min-width: 930px) {
+  #shareList {
+    height: calc(100vh - 9.8em);
+    overflow-y: auto;
+  }
+}
+</style>